requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
ansicon==1.89.0
anyio==3.6.2
arrow==1.2.3
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.3.0
blessed==1.20.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
croniter==1.3.14
dateutils==0.6.12
deepdiff==6.3.0
fastapi==0.88.0
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.4.0
h11==0.14.0
idna==3.4
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
jinxed==1.2.0
lightning==2.0.2
lightning-cloud==0.5.34
lightning-utilities==0.8.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.3
ordered-set==4.1.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.0
psutil==5.9.5
pydantic==1.10.7
Pygments==2.15.1
PyJWT==2.6.0
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.2
pytz==2023.3
PyYAML==6.0
readchar==4.0.5
requests==2.30.0
rich==13.3.5
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.22.0
starsessions==1.3.0
sympy==1.11.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
wcwidth==0.2.6
websocket-client==1.5.1
websockets==11.0.2
yarl==1.9.2